Improved CouchDB database security
The 'newebe' database now has an associated member, which is a user. All
access to this database will have to be done through this user (or
through the CouchDB server admin)

This commit creates this user in the database with the CouchDB server
admin. Both are to be specified in local_settings.py, or in settings.py